Added `ser/values` module
